Merge pull request #544 from O2-Bionics/feature/543-community-service-create-skeleton

+++ src/Services/c-gen/O2NextGen.CertificateManagement.Domain/UseCases/ForCategory/DeleteCategory/DeleteCategoryCommandHandler.cs
@@ -1,17 +1,18 @@
 using MediatR;
 using O2NextGen.CertificateManagement.Domain.Data;
 using O2NextGen.CertificateManagement.Domain.Data.Queries;
+using O2NextGen.CertificateManagement.Domain.Entities;
 
 namespace O2NextGen.CertificateManagement.Domain.UseCases.ForCategory.DeleteCategory;
 
 public class DeleteCategoryCommandHandler : IRequestHandler<DeleteCategoryCommand, Unit>
 {
-    private readonly IQueryHandler<CategoryQuery, Entities.Category> _userGroupQueryHandler;
-    private readonly IRepository<Entities.Category> groupsRepository;
+    private readonly IQueryHandler<CategoryQuery, Category> _userGroupQueryHandler;
+    private readonly IRepository<Category> groupsRepository;
 
     public DeleteCategoryCommandHandler(
-        IQueryHandler<CategoryQuery, Entities.Category> userGroupQueryHandler,
-        IRepository<Entities.Category> groupsRepository)
+        IQueryHandler<CategoryQuery, Category> userGroupQueryHandler,
+        IRepository<Category> groupsRepository)
     {
         _userGroupQueryHandler =
             userGroupQueryHandler ?? throw new ArgumentNullException(nameof(userGroupQueryHandler));
@@ -24,10 +25,7 @@ public class DeleteCategoryCommandHandler : IRequestHandler<DeleteCategoryComman
             new CategoryQuery(request.Id),
             cancellationToken);
 
-        if (category is object)
-        {
-            await groupsRepository.DeleteAsync(category, cancellationToken);
-        }
+        if (category is object) await groupsRepository.DeleteAsync(category, cancellationToken);
 
         return Unit.Value;
     }
